smyrna set_color_button_widget: squash -Wconversion warnings
The values on the RHSs here are only ever within range of the target type.

+++ cmd/smyrna/gui/topviewsettings.c
@@ -79,9 +79,9 @@ static int set_color_button_widget(char *attrib, char *widget_name)
     }
     if (buf) {
 	colorxlate(buf, &cl, RGBA_DOUBLE);
-	color.red = (int) (cl.u.RGBA[0] * 65535.0);
-	color.green = (int) (cl.u.RGBA[1] * 65535.0);
-	color.blue = (int) (cl.u.RGBA[2] * 65535.0);
+	color.red = (guint16)(cl.u.RGBA[0] * 65535.0);
+	color.green = (guint16)(cl.u.RGBA[1] * 65535.0);
+	color.blue = (guint16)(cl.u.RGBA[2] * 65535.0);
 	gtk_color_button_set_color((GtkColorButton *)
 				   glade_xml_get_widget(xml, widget_name),
 				   &color);
